Bug 1171: Two minor 64-bit Aips/Casa issues

+++ branches/CEP-BB-Bug1158-task-branch/CEP/BB/BBSKernel/include/BBSKernel/Prediffer.h
@@ -160,7 +160,7 @@ private:
         unsigned long long      visCount;
 
         vector<casa::LSQFit*>   equations;
-        vector<size_t>          coeffIndex;
+        vector<uint>            coeffIndex;
         vector<const double*>   perturbedRe, perturbedIm;
         vector<double>          partialRe, partialIm;
     };
